Add GUI material editor window
Meshes no longer have material attributes - there is now a MaterialMesh class which holds a mesh and a material instead. This is so that meshes do not have to be recreated when creating entities with the same mesh, but different materials.

PBR materials now contain both textures and numerical values for albedo, metallicness and roughness. The user can choose whether to use a texture or the numerical value from the GUI.

Other changes:
- The camera now has a smooth movement option
- Many more camera settings are now editable within the GUI (including movement speed, mouse sensitivity, deceleration, and yaw/pitch)
- Entities now have a name attribute which is used in the GUI
- The user can now choose whether tone mapping is used