Use sky as environment map for IBL
The cubemap does not quite look right - the faces do not line up correctly. Using an FOV of PI/3.5 radians seems to be the most accurate for the side faces, but the top and bottom faces do not fit as well. However, the results should still be acceptable for materials that are not perfectly reflective (i.e. those with non-zero roughness).

Other changes/additions:
- Add fast irradiance computation option
- Refactor IBL texture rendering into a separate method so that textures can be rendered without having to re-initialise all the shaders
- Refactor shader file reading into the relevant methods so that the code does not have to be included in the argument